JAMES-1877 Extract delivery to single server to MailDelivrer
